[API BREAK] Adds a dataset capability whether supports active flag
Dataset may or may not active flag capability. If it supports the active flag concept, the MDAL_D_data
can return a proper values of active flag for faces. If it does not support it (e.g. data is defined on
faces, so active flag does not have meaning), the MDAL_D_data will fail.